Kitty-grade extras, cursor trail animation, selection fixes & speed work
Customization:
- shell=zsh -l spawns any command; env NAME=VALUE exports child variables
- cursor_text_color, cursor_stop_blinking_after (solid after idle), bar and
  underline shapes fully wired
- visual_bell_duration: BEL now flashes the screen with a fading foreground
  overlay (previously silently ignored)
- adjust_line_height / adjust_column_width / adjust_baseline pixel tuning
- cursor_trail: kitty-style animated cursor — the block glides to its new
  position with ease-out, a fading smear and the character riding along;
  retargets mid-flight, snaps when scrolled/hidden, ticks at ~80fps only
  while animating (zero cost when idle)

Selection semantics (kitty-like):
- drag threshold: a click or pixel jitter never selects, only a real drag
- single click clears an existing selection
- typing and key repeat dismiss the selection; Ctrl+Shift+C still copies

Performance & size (verified pixel-identical):
- all per-pixel blend loops use the exact (v+1)*257>>16 division-by-255
  identity — no divisions in the hottest paths
- Cell shrunk 24 -> 16 bytes (scrollback memory -33%); Glyph dead field
  removed (~300KB less cache)
- default build now -O2 and stripped: 233KB binary (was 638KB debug build)

Docs:
- new CUSTOMIZING.md: guided tour with ready-to-copy example configs;
  README/CONFIG.md cross-linked to it

All suites green (57 parser + 15 real-app PTY tests, ASan/UBSan/LSan clean),
all three build configurations clean.

+++ include/vt_parser.h
@@ -36,9 +36,9 @@ typedef struct {
     uint32_t char_code;
     uint32_t fg_color;
     uint32_t bg_color;
-    int wrapped;
+    uint8_t wrapped;
     uint8_t attrs;
-} Cell;
+} Cell; // 16 bytes — keep it tight: scrollback holds rows*cols of these
 
 typedef struct KittyImage {
     int id;
